+++ tools/src/create/run.rs
@@ -72,7 +72,7 @@ self.addEventListener('activate', (e) => {
 });
 
 const r = router();
-r.register('/*', new CacheFirst(SHELL));
+r.register('*', new CacheFirst(SHELL));
 r.register('/api/*', new NetworkFirst(API, { timeout: 3000 }));
 
 self.addEventListener('fetch', (e) => {